Add Series Domain
This adds a pattern called Repository Pattern where there is a
layer that accesses the Datastore for us. This allows us to
change the implementation at any time (_for tests for instance_)
without having to worry about the logic.

A Domain's Repository is its way to access its specific data. If
there are Aggregates, as in 2+ groupings of Domains, that should
be a layer ABOVE the Domains. Aggregates are "smarts" attached to
a Domain and shouldn't be encoded _inside_ the Domain itself.

If we do go down that path, we should have Aggregate Domains and
not muddy the Repositories with those details